refactor(config): migrate from constants to config for site metadata
