+++ src/app/manifest.ts
@@ -1,5 +1,5 @@
 // app/manifest.ts
-import { siteConfig } from '@/lib/constants';
+import { siteConfig } from '@/lib/config';
 import type { MetadataRoute } from 'next';
 
 export default function manifest(): MetadataRoute.Manifest {
