+++ src/app/sitemap.ts
@@ -1,4 +1,4 @@
-import { siteConfig } from '@/lib/constants';
+import { siteConfig } from '@/lib/config';
 import type { MetadataRoute } from 'next';
 
 export default function sitemap(): MetadataRoute.Sitemap {
